# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Audit Brief API",
    "version": "0.2.1",
...
    "contact": {
      "name": "Audit Brief Platform Team"
    },
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "https://uat.uno.wcgt.in/auditbrief",
      "description": "UAT / production-equivalent (basePath `/auditbrief`)."
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/bookmarks": {
      "get": {
        "tags": ["Bookmarks"],
        "operationId": "listBookmarks",
        "summary": "List the caller's bookmarks",
        "description": "Authenticated. Can filter by `auditBriefId` or `episodeId`.",
        "parameters": [
          {
            "in": "query",
            "name": "auditBriefId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Return only bookmarks on this audit brief."
          },
          {
            "in": "query",
            "name": "episodeId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Return only bookmarks on this learning-path episode."
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated bookmarks for the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Bookmark"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Bookmarks"],
        "operationId": "createBookmark",
        "summary": "Create a bookmark",
        "description": "Authenticated. Exactly one of `auditBriefId` or `episodeId` must be\nsupplied. The referenced parent must exist, otherwise `400`.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookmarkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bookmark created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Bookmark"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/bookmarks/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/IdPathParam"
        }
      ],
      "put": {
        "tags": ["Bookmarks"],
        "operationId": "updateBookmark",
        "summary": "Update a bookmark (note only)",
        "description": "Authenticated, owner-scoped. Only `note` may be updated.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookmarkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated bookmark.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Bookmark"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Caller does not own this bookmark.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": ["Bookmarks"],
        "operationId": "deleteBookmark",
        "summary": "Delete a bookmark",
        "description": "Authenticated, owner-scoped.",
        "responses": {
          "200": {
            "description": "Bookmark deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bookmark deleted"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Caller does not own this bookmark.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/activity": {
      "post": {
        "tags": ["Activity"],
        "operationId": "logActivity",
        "summary": "Log a user activity event (fire-and-forget)",
        "description": "Authenticated. Records a single `UserActivity` row. The response body\nis `null` — the client fires-and-forgets. This feeds `/api/admin/analytics`.\n\nNote: `auditBriefId` / `episodeId` / `graphId` are accepted as free\nstrings (not validated as UUIDs at the schema level); the server will\nnull them out at persistence time rather than reject.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityEventRequest"
              },
              "example": {
                "activityType": "listen",
                "auditBriefId": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10",
                "metadata": {
                  "seconds": 42
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity recorded. Response body is the JSON literal `null`\n(OpenAPI 3.0.3 has no `null` type, so no body schema is declared).\n"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth v4 JWT session cookie. `httpOnly`, `SameSite=Lax`, and `Secure`\nin production. Over HTTPS the cookie is named\n`__Secure-next-auth.session-token`.\n\nIssued by `POST /api/auth/callback/credentials` (dev) or the Azure AD\nOAuth flow (`GET /api/auth/signin/azure-ad` → provider → `GET /api/auth/callback/azure-ad`).\nContains a signed JWT with `userId`, `role`, `email`, and `jti` (for\nserver-side revocation).\n"
      }
    },
...
    "parameters": {
      "IdPathParam": {
        "in": "path",
        "name": "id",
        "required": true,
        "description": "Canonical UUID (8-4-4-4-12 hex).",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "example": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"
      },
      "PageParam": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "1-based page index."
      },
      "LimitParam": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Page size (max 100)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request, missing required fields, or invalid values.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 400,
              "error_code": "BAD_REQUEST",
              "message": "key parameter required",
              "request_id": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "No session cookie or the session has expired / been revoked.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 401,
              "error_code": "UNAUTHORIZED",
              "message": "Authentication required",
              "request_id": "3f2e7a6d-4b1a-4f9d-8a77-1b4c8e6d2a10"
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Resource does not exist (or is archived where archived rows are hidden).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 404,
              "error_code": "NOT_FOUND",
              "message": "AuditBrief not found"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request body failed Zod validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 422,
              "error_code": "VALIDATION_FAILED",
              "message": "Validation failed",
              "details": [
                {
                  "path": ["password"],
                  "message": "Password must be at least 12 characters",
                  "code": "too_small"
                }
              ]
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unhandled server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 500,
              "error_code": "INTERNAL_ERROR",
              "message": "Internal server error"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["status", "error_code", "message"],
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 422
          },
          "error_code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_FAILED",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error classification."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message. Safe to surface to end-users."
          },
          "details": {
            "description": "Free-form diagnostic detail. For `VALIDATION_FAILED` this is a Zod\n`issues` array; for other codes it is commonly absent. May be null.\n"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "description": "Echo of the `x-request-id` header — use for log correlation."
          }
        }
      },
      "PaginatedEnvelope": {
        "type": "object",
        "required": ["data", "pagination"],
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          },
          "pagination": {
            "type": "object",
            "required": ["page", "limit", "total", "total_pages"],
            "properties": {
              "page": {
                "type": "integer",
                "minimum": 1,
                "example": 1
              },
              "limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100,
                "example": 20
              },
              "total": {
                "type": "integer",
                "minimum": 0,
                "example": 137
              },
              "total_pages": {
                "type": "integer",
                "minimum": 0,
                "example": 7
              }
            }
          }
        }
      },
...
      "ActivityType": {
        "type": "string",
        "enum": [
          "listen",
          "bookmark",
          "unbookmark",
          "complete_episode",
          "view_audit_brief",
          "view_path",
          "search",
          "favorite",
          "unfavorite",
          "signin",
          "signout",
          "signin_failed"
        ]
      },
...
      "CreateBookmarkRequest": {
        "type": "object",
        "required": ["timestampSeconds"],
        "description": "Exactly one of `auditBriefId` or `episodeId` is required (enforced server-side).",
        "properties": {
          "auditBriefId": {
            "type": "string",
            "format": "uuid"
          },
          "episodeId": {
            "type": "string",
            "format": "uuid"
          },
          "timestampSeconds": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
      "UpdateBookmarkRequest": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
      "Bookmark": {
        "type": "object",
        "required": ["id", "userId", "timestampSeconds", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "auditBriefId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "episodeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "timestampSeconds": {
            "type": "number",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ActivityEventRequest": {
        "type": "object",
        "required": ["activityType"],
        "properties": {
          "activityType": {
            "$ref": "#/components/schemas/ActivityType"
          },
          "auditBriefId": {
            "type": "string",
            "description": "Not UUID-validated at schema level."
          },
          "episodeId": {
            "type": "string"
          },
          "graphId": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "description": "Free-form structured payload (e.g. playback position, search query).",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}